work on quick tips
