+++ app/assets/javascripts/discourse/app/components/search-menu.hbs
@@ -25,6 +25,7 @@
       @searchTopics={{this.includesTopics}}
       @typeFilter={{this.typeFilter}}
       @updateTypeFilter={{this.updateTypeFilter}}
+      @updateTerm={{this.updateTerm}}
     />
   </MenuPanel>
 </div>
