+++ app/assets/javascripts/discourse/app/components/search-menu/results/initial-options.hbs
@@ -25,7 +25,10 @@
         />
       {{/if}}
     {{else}}
-      <SearchMenu::Results::RandomQuickTip />
+      <SearchMenu::Results::RandomQuickTip
+        @updateTerm={{@updateTerm}}
+        @focusSearchInput={{@focusSearchInput}}
+      />
 
       {{#if (and this.currentUser this.siteSettings.log_search_queries)}}
         {{! search-menu-recent-searches }}
