+++ app/assets/javascripts/discourse/app/components/search-menu/results/type/category.hbs
@@ -1,3 +1,1 @@
-<a href={{@result.url}} class="search-link">
-  {{category-link @result link=false}}
-</a>
+{{category-link @result link=false}}
