+++ app/assets/javascripts/discourse/app/components/search-menu/results/type/post.hbs
@@ -1,4 +1,2 @@
-<a href={{@result.url}} class="search-link">
-  {{i18n "search.post_format" @result}}
-  <SearchMenu::Results::Blurb @result={{@result}} @term={{@term}} />
-</a>
+{{i18n "search.post_format" @result}}
+<SearchMenu::Results::Blurb @result={{@result}} @term={{@term}} />
