+++ app/assets/javascripts/discourse/app/components/search-menu/results/type/tag.hbs
@@ -1,4 +1,2 @@
-<a href={{@result.url}} class="search-link">
-  {{d-icon "tag"}}
-  {{discourse-tag @result.id tagName="span"}}
-</a>
+{{d-icon "tag"}}
+{{discourse-tag @result.id tagName="span"}}
