+++ app/assets/javascripts/discourse/app/components/search-menu/results/type/topic.hbs
@@ -1,24 +1,22 @@
-<a href={{@result.url}} class="search-link">
-  <span class="topic">
-    <span class="first-line">
-      <TopicStatus @topic={{@result.topic}} @disableActions={{true}} />
-      <span class="topic-title" data-topic-id={{@result.topic.id}}>
-        {{#if this.unescapeHeadline}}
-          <span>{{this.unescapedHeadline}}</span>
-        {{else}}
-          <SearchMenu::HighlightedSearch
-            @string={{@result.topic.fancyTitle}}
-            @term={{@term}}
-          />
-        {{/if}}
-      </span>
-    </span>
-    <span class="second-line">
-      {{category-link @result.topic.category link=false}}
-      {{#if this.siteSettings.tagging_enabled}}
-        {{discourse-tags @result.topic tagName="span"}}
+<span class="topic">
+  <span class="first-line">
+    <TopicStatus @topic={{@result.topic}} @disableActions={{true}} />
+    <span class="topic-title" data-topic-id={{@result.topic.id}}>
+      {{#if this.unescapeHeadline}}
+        <span>{{this.unescapedHeadline}}</span>
+      {{else}}
+        <SearchMenu::HighlightedSearch
+          @string={{@result.topic.fancyTitle}}
+          @term={{@term}}
+        />
       {{/if}}
     </span>
   </span>
-  <SearchMenu::Results::Blurb @result={{@result}} @term={{@term}} />
-</a>
+  <span class="second-line">
+    {{category-link @result.topic.category link=false}}
+    {{#if this.siteSettings.tagging_enabled}}
+      {{discourse-tags @result.topic tagName="span"}}
+    {{/if}}
+  </span>
+</span>
+<SearchMenu::Results::Blurb @result={{@result}} @term={{@term}} />
