+++ app/assets/javascripts/discourse/app/components/search-menu/search-term.hbs
@@ -1,7 +1,7 @@
 <Input
   id="search-term"
   @type="text"
-  @value={{@value}}
+  @value={{@term}}
   autocomplete="off"
   placeholder={{I18n "search.title"}}
   aria-label={{I18n "search.title"}}
